Revert "More aditional shooter fixes"
This reverts commit 56cfe237e72d86117aad3d483232a096cfe765a4.

+++ src/main/java/frc/robot/commands/ShootWaitVelocity.java
@@ -40,7 +40,7 @@ public class ShootWaitVelocity extends CommandBase {
   @Override
   public void execute() {
    
-    double velocity = m_subsystem2.getVelocityError();
+    int velocity = m_subsystem2.getVelocityError();
 
     if ((Math.abs(velocity) < 300)){
